# docs/endpoints/openapi-systems-ops.yaml
openapi: 3.0.3
info:
  title: Atlas API — Systems Ops
  version: '1.0.0'
  description: |
    Contrato OpenAPI ULTRA DETALLADO, generado leyendo directamente el código fuente
    (controladores, schemas Zod, mappers, servicios y repositorios), del módulo
    `systems-ops` del backend Atlas: catálogo técnico de endpoints/herramientas/entidades
    de datos, log de acciones (auditoría HTTP), cola de revisión de gobierno de datos,
    perfiles y runs de stress testing, y suites/pasos/runs de pruebas E2E internas.

    ## Envolturas globales (aplican a TODOS los endpoints de este documento)

    Este backend registra dos providers globales en `src/app.module.ts`
    (`APP_INTERCEPTOR` y `APP_FILTER`) que envuelven **toda** respuesta, éxito o error.
    Los schemas de `data` documentados endpoint por endpoint abajo son el contenido
    de la propiedad `data`, **no** el body HTTP completo.

    **Éxito** (`ResponseInterceptor`, `src/common/interceptors/response.interceptor.ts`):
    ```json
    { "requestId": "corr-id-o-undefined", "data": { "...": "..." }, "timestamp": "2026-07-07T12:00:00.000Z" }
    ```

    **Error** (`HttpExceptionFilter`, `src/common/filters/http-exception.filter.ts`):
    ```json
    { "requestId": "corr-id-o-undefined", "error": { "code": "VALIDATION_ERROR", "message": "..." }, "timestamp": "2026-07-07T12:00:00.000Z" }
    ```
    `error.code` se deriva del status HTTP: `400→VALIDATION_ERROR`, `401→UNAUTHORIZED`,
...
    de query/params/body.

...

...
    ```ts
...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /systems/test-suites:
    post:
      operationId: createTestSuite
      tags: [SystemsTest]
      summary: Crear una suite de pruebas para un módulo del backend
      description: Usa `@CurrentUser()` como `createdBy`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTestSuiteRequest' }
      responses:
        '201':
          description: Suite creada (sin pasos).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TestSuiteDetail'
        '400':
          description: |
            Uno de:
            - `VALIDATION_ERROR` (Zod).
            - `PRODUCTION_READONLY_REQUIRES_SAFE_SUITE` (`environmentScope` incluye `PRODUCTION_READONLY` sin `isSafeForProduction=true`).
            - `SYSTEM_TEST_SUITE_CREATE_FAILED` (otro error de persistencia).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: SYSTEM_TEST_SUITE_OR_STEP_ALREADY_EXISTS (código duplicado)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    get:
      operationId: listTestSuites
      tags: [SystemsTest]
      summary: Listar suites de prueba con filtros
      parameters:
        - $ref: '#/components/parameters/ModuleQuery'
        - name: suiteType
          in: query
          schema: { type: string, minLength: 1, maxLength: 200 }
          description: 'Lista de tipos separada por coma, p. ej. `INTEGRATION,SMOKE`.'
        - name: enabled
          in: query
          schema: { type: boolean }
          description: Filtra por `isEnabled`.
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Lista paginada de suites.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TestSuiteList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/test-suites/{suiteId}:
    get:
      operationId: getTestSuite
      tags: [SystemsTest]
      summary: Obtener una suite con todos sus pasos ordenados
      parameters:
        - $ref: '#/components/parameters/SuiteIdPath'
      responses:
        '200':
          description: Suite con sus pasos (orden stepOrder ASC).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TestSuiteDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_TEST_SUITE_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    patch:
      operationId: updateTestSuite
      tags: [SystemsTest]
      summary: Actualizar campos de una suite existente (parcial)
      description: 'Mismos campos que `CreateTestSuiteRequest`, todos opcionales — pero el body debe traer al menos uno (`400` con mensaje "Debe enviar al menos un campo para actualizar." si viene vacío).'
      parameters:
        - $ref: '#/components/parameters/SuiteIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTestSuiteRequest' }
      responses:
        '200':
          description: Suite actualizada (pasos sin cambios por este endpoint).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TestSuiteDetail'
        '400':
          description: |
            Uno de: `VALIDATION_ERROR`, `PRODUCTION_READONLY_REQUIRES_SAFE_SUITE`,
            `SYSTEM_TEST_SUITE_UPDATE_FAILED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_TEST_SUITE_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '409':
          description: SYSTEM_TEST_SUITE_OR_STEP_ALREADY_EXISTS
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/test-suites/{suiteId}/steps:
    post:
      operationId: createTestStep
      tags: [SystemsTest]
      summary: Agregar un paso (llamada HTTP) a una suite de pruebas
      parameters:
        - $ref: '#/components/parameters/SuiteIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTestStepRequest' }
      responses:
        '201':
          description: Paso creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TestStep'
        '400':
          description: |
            Uno de:
            - `VALIDATION_ERROR`.
            - `SYSTEM_TEST_STEP_ENDPOINT_NOT_FOUND` (`endpointId` enviado no existe en catálogo).
            - `PRODUCTION_SAFE_SUITE_ONLY_ALLOWS_READONLY_METHODS` (suite `isSafeForProduction=true` y `method` no es GET/HEAD/OPTIONS).
            - `PRODUCTION_SAFE_SUITE_PATH_LOOKS_MUTATING` (suite `isSafeForProduction=true` y `pathTemplate` matchea patrón de mutación).
            - `SYSTEM_TEST_STEP_CREATE_FAILED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_TEST_SUITE_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '409':
          description: SYSTEM_TEST_SUITE_OR_STEP_ALREADY_EXISTS
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/test-suites/{suiteId}/steps/{stepId}:
    patch:
      operationId: updateTestStep
      tags: [SystemsTest]
      summary: Actualizar un paso existente de una suite (parcial)
      parameters:
        - $ref: '#/components/parameters/SuiteIdPath'
        - name: stepId
          in: path
          required: true
          schema: { type: string, pattern: '^[1-9][0-9]*$' }
          description: Id numérico del paso.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTestStepRequest' }
      responses:
        '200':
          description: Paso actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TestStep'
        '400':
          description: |
            Uno de: `VALIDATION_ERROR` (incluye body vacío), `SYSTEM_TEST_STEP_ENDPOINT_NOT_FOUND`,
            `PRODUCTION_SAFE_SUITE_ONLY_ALLOWS_READONLY_METHODS`, `PRODUCTION_SAFE_SUITE_PATH_LOOKS_MUTATING`,
            `SYSTEM_TEST_STEP_UPDATE_FAILED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_TEST_SUITE_NOT_FOUND o SYSTEM_TEST_STEP_NOT_FOUND (paso no existe o no pertenece a la suite)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '409':
          description: SYSTEM_TEST_SUITE_OR_STEP_ALREADY_EXISTS
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/test-suites/{suiteId}/steps/reorder:
    post:
      operationId: reorderTestSteps
      tags: [SystemsTest]
      summary: Reordenar en bloque los pasos de una suite
      parameters:
        - $ref: '#/components/parameters/SuiteIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReorderTestStepsRequest' }
      responses:
        '201':
          description: Pasos reordenados (todos los pasos de la suite tras el reorden, orden stepOrder ASC).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TestStepList'
        '400':
          description: |
            Uno de:
            - `VALIDATION_ERROR`.
            - `SYSTEM_TEST_STEP_NOT_IN_SUITE:<ids>` (algún `stepId` enviado no pertenece a la suite; el mensaje incluye los ids desconocidos).
            - `SYSTEM_TEST_STEP_DUPLICATED_ORDER` (`stepOrder` duplicado entre los items enviados).
            - `SYSTEM_TEST_STEP_REORDER_FAILED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_TEST_SUITE_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/test-suites/{suiteId}/run:
    post:
      operationId: runTestSuite
      tags: [SystemsTest]
      summary: Ejecutar una suite de pruebas (dry-run simulado o HTTP real), paso a paso
      description: |
        Evalúa aserciones y encadena contexto entre pasos (extractors). Usa `@CurrentUser()`
        como `triggeredBy`. Los errores dentro de un paso individual (p. ej. `pathTemplate` no
        resuelve, timeout HTTP) **no** se propagan como excepción HTTP: el paso queda `FAILED`
        con `errorMessage` y la suite continúa o se detiene según `continueOnFailure` de cada paso.
      parameters:
        - $ref: '#/components/parameters/SuiteIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunTestSuiteRequest' }
      responses:
        '201':
          description: Resultado agregado del run + detalle de cada paso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TestRunDetail'
        '400':
          description: |
            Uno de:
            - `VALIDATION_ERROR`.
            - `SYSTEM_TEST_SUITE_DISABLED` (`isEnabled=false`).
            - `SYSTEM_TEST_BASE_URL_REQUIRED_FOR_REAL_RUN` (`dryRun=false` sin `baseUrl`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: |
            Uno de:
            - Rol fuera de la lista permitida (mensaje genérico de `RolesGuard`).
            - `SYSTEM_TEST_ENVIRONMENT_NOT_ALLOWED_FOR_SUITE` (`environment` no está en `environmentScope` de la suite).
            - `SYSTEM_TEST_PRODUCTION_EXECUTION_BLOCKED` (`environment=PRODUCTION_READONLY` y `isSafeForProduction=false` o `dryRun=false`).
            - `SYSTEM_TEST_LOCAL_ENVIRONMENT_REQUIRES_LOCAL_BASE_URL` (ejecución real en `LOCAL` con `baseUrl` cuyo host no es localhost/127.0.0.1/host.docker.internal).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '404':
          description: SYSTEM_TEST_SUITE_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /systems/test-runs:
    get:
      operationId: listTestRuns
      tags: [SystemsTest]
      summary: Listar ejecuciones de suites de prueba
      parameters:
        - $ref: '#/components/parameters/SuiteIdQuery'
        - name: status
          in: query
          schema: { type: string, enum: [QUEUED, RUNNING, PASSED, FAILED, CANCELLED] }
        - name: environment
          in: query
          schema: { type: string, enum: [LOCAL, STAGING, PRODUCTION_READONLY] }
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Lista paginada de runs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TestRunList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /systems/test-runs/{runId}:
    get:
      operationId: getTestRun
      tags: [SystemsTest]
      summary: Detalle de una ejecución de suite, incluyendo el resultado de cada paso
      parameters:
        - $ref: '#/components/parameters/RunIdPath'
      responses:
        '200':
          description: Run con sus pasos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_TestRunDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: SYSTEM_TEST_RUN_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT emitido por `POST /api/v1/internal/auth/login` (o `/api/v1/auth/login` para
        roles no internos). Debe pertenecer a un usuario con rol en:
        system_admin, platform_admin, admin, qa_engineer, devops, risk_analyst,
        compliance_analyst, readonly_auditor.
  responses:
    BadRequest:
      description: Entrada inválida (Zod) o regla de negocio violada — ver detalle por endpoint.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Falta `Authorization`, formato inválido, token expirado/inválido o revocado.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: 'Rol autenticado no está en la lista permitida: "El usuario autenticado no tiene permiso para esta operación."'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  parameters:
    PageQuery:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: Número de página (entero positivo, coercionado desde string).
    LimitQuery:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Tamaño de página (máx 100).
    ModuleQuery:
      name: module
      in: query
      schema: { type: string, minLength: 1, maxLength: 120 }
      description: Filtra por módulo.
...
    SuiteIdQuery:
      name: suiteId
      in: query
      schema: { type: string, pattern: '^[1-9][0-9]*$' }
      description: Filtra por suite.
...
    SuiteIdPath:
      name: suiteId
      in: path
      required: true
      schema: { type: string, pattern: '^[1-9][0-9]*$' }
      description: Id numérico de la suite de pruebas.
    RunIdPath:
      name: runId
      in: path
      required: true
      schema: { type: string, pattern: '^[1-9][0-9]*$' }
      description: Id numérico del run de suite.
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        requestId: { type: string, nullable: true, example: 'req-8f3c2a' }
        error:
          type: object
          properties:
            code:
              type: string
              enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, INTERNAL_ERROR]
            message: { type: string, example: 'SYSTEM_ENDPOINT_NOT_FOUND' }
          required: [code, message]
        timestamp: { type: string, format: date-time }
      required: [error, timestamp]
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
      required: [page, limit, total, totalPages]

    # ---------- Entidades del dominio (forma real devuelta por los mappers) ----------
...
    TestSuite:
      type: object
      description: Forma producida por `mapTestSuite`.
      properties:
        suiteId: { type: string }
        code: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        module: { type: string }
        suiteType: { type: string, enum: [INTEGRATION, SMOKE, REGRESSION, E2E_API, LOAD] }
        executionMode: { type: string, example: SYNC_OR_JOB }
        environmentScope: { type: array, items: { type: string, enum: [LOCAL, STAGING, PRODUCTION_READONLY] } }
        isEnabled: { type: boolean }
        requiresSeedData: { type: boolean }
        isSafeForProduction: { type: boolean }
        requiresDestructivePermission: { type: boolean }
    TestStep:
      type: object
      description: Forma producida por `mapTestStep`.
      properties:
        stepId: { type: string }
        suiteId: { type: string }
        endpointId: { type: string, nullable: true }
        stepOrder: { type: integer }
        name: { type: string }
        inputMode: { type: string, enum: [DEFAULT, CONFIGURABLE, GENERATED, FROM_PREVIOUS_STEP] }
        method: { type: string, enum: [GET, POST, PUT, PATCH, DELETE, OPTIONS, HEAD] }
        pathTemplate: { type: string, example: '/systems/endpoints/{{endpointId}}' }
        defaultHeaders: { type: object }
        defaultPayload: { type: object }
        configSchema: { type: object }
        extractors: { type: object }
        assertions: { type: object, example: { expectedStatusCodes: [200, 201] } }
        continueOnFailure: { type: boolean }
        cleanupRequired: { type: boolean }
...
    TestStepRun:
      type: object
      description: Forma producida por `mapTestStepRun`.
      properties:
        stepRunId: { type: string }
        testRunId: { type: string }
        stepId: { type: string }
        status: { type: string, enum: [PASSED, FAILED, SKIPPED] }
        requestPayloadSanitized: { type: object }
        responseBodySanitized: { type: object }
        statusCode: { type: integer, nullable: true }
        durationMs: { type: number }
        errorMessage: { type: string, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }

    # ---------- Request bodies ----------
...
    CreateTestSuiteRequest:
      type: object
      properties:
        code: { type: string, minLength: 3, maxLength: 180, pattern: '^[A-Z0-9_]+$' }
        name: { type: string, minLength: 3, maxLength: 220 }
        description: { type: string, maxLength: 4000 }
        module: { type: string, minLength: 2, maxLength: 120 }
        suiteType: { type: string, enum: [INTEGRATION, SMOKE, REGRESSION, E2E_API, LOAD], default: INTEGRATION }
        environmentScope:
          type: array
          items: { type: string, enum: [LOCAL, STAGING, PRODUCTION_READONLY] }
          minItems: 1
          maxItems: 3
          default: [LOCAL, STAGING]
        isEnabled: { type: boolean, default: true }
        requiresSeedData: { type: boolean, default: true }
        isSafeForProduction: { type: boolean, default: false }
        requiresDestructivePermission:
          type: boolean
          description: 'Si se omite, se calcula como !isSafeForProduction.'
      required: [code, name, module]
    UpdateTestSuiteRequest:
      allOf:
        - $ref: '#/components/schemas/CreateTestSuiteRequest'
      description: 'Todos los campos de CreateTestSuiteRequest, opcionales (.partial()) — el body debe traer al menos uno.'
    CreateTestStepRequest:
      type: object
      properties:
        endpointId:
          type: string
          pattern: '^[1-9][0-9]*$'
          nullable: true
          description: Debe existir en el catálogo si se envía.
        stepOrder: { type: integer, minimum: 1, maximum: 500 }
        name: { type: string, minLength: 3, maxLength: 220 }
        inputMode: { type: string, enum: [DEFAULT, CONFIGURABLE, GENERATED, FROM_PREVIOUS_STEP], default: DEFAULT }
        method: { type: string, enum: [GET, POST, PUT, PATCH, DELETE, OPTIONS, HEAD] }
        pathTemplate:
          type: string
          minLength: 1
          maxLength: 1200
          description: 'Debe iniciar con "/" o "http(s)://". Soporta placeholders {{...}}.'
        defaultHeaders: { type: object, default: {} }
        defaultPayload: { type: object, default: {} }
        configSchema: { type: object, default: {} }
        extractors: { type: object, default: {}, description: 'JSONPath para extraer valores de la respuesta hacia el contexto de la suite.' }
        assertions: { type: object, default: { expectedStatusCodes: [200, 201] } }
        continueOnFailure: { type: boolean, default: false }
        cleanupRequired: { type: boolean, default: false }
      required: [stepOrder, name, method, pathTemplate]
    UpdateTestStepRequest:
      allOf:
        - $ref: '#/components/schemas/CreateTestStepRequest'
      description: 'Todos los campos de CreateTestStepRequest, opcionales (.partial()) — el body debe traer al menos uno.'
    ReorderTestStepsRequest:
      type: object
      properties:
        steps:
          type: array
          minItems: 1
          maxItems: 500
          items:
            type: object
            properties:
              stepId: { type: string, pattern: '^[1-9][0-9]*$' }
              stepOrder: { type: integer, minimum: 1, maximum: 500 }
            required: [stepId, stepOrder]
      required: [steps]
    RunTestSuiteRequest:
      type: object
      properties:
        environment: { type: string, enum: [LOCAL, STAGING, PRODUCTION_READONLY], default: LOCAL }
        dryRun: { type: boolean, default: true, description: 'Si true (o si no se envía baseUrl), los pasos se resuelven sin llamada HTTP real.' }
        baseUrl: { type: string, format: uri, description: Requerido si dryRun=false. }
        config: { type: object, default: {}, description: 'Config por paso (step_<n> o nombre del paso) usada para resolver templates.' }
        headers:
          type: object
          additionalProperties: { type: string }
          default: {}
        timeoutMs: { type: integer, minimum: 100, maximum: 60000, default: 10000 }

    # ---------- Result payloads compuestos (no mapeados 1:1 a una entidad) ----------
...
    TestSuiteDetailResult:
      type: object
      properties:
        suite: { $ref: '#/components/schemas/TestSuite' }
        steps: { type: array, items: { $ref: '#/components/schemas/TestStep' } }
    TestRunDetailResult:
      type: object
      properties:
        run: { $ref: '#/components/schemas/TestRun' }
        steps: { type: array, items: { $ref: '#/components/schemas/TestStepRun' } }

    # ---------- Envelopes de respuesta (Envelope global + data específico) ----------
...
    Envelope_TestSuiteList:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/TestSuite' } }
            meta: { $ref: '#/components/schemas/PaginationMeta' }
    Envelope_TestSuiteDetail:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/TestSuiteDetailResult' }
    Envelope_TestStep:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/TestStep' }
    Envelope_TestStepList:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/TestStep' } }
    Envelope_TestRunList:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data:
          type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/TestRun' } }
            meta: { $ref: '#/components/schemas/PaginationMeta' }
    Envelope_TestRunDetail:
      type: object
      properties:
        requestId: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }
        data: { $ref: '#/components/schemas/TestRunDetailResult' }